refactor: pieceResponse 패키지 변경
